+++ common/src/index.ts
@@ -57,5 +57,6 @@ export * from './lib/json/ImageViewData';
 export * from './lib/json/ViewerEvent';
 export * from './lib/json/LatestVersion';
 export * from './lib/json/AddContentSourceRequest';
+export * from './lib/json/WatermarkSettings';
 export * from './lib/scrape/ScraperHelpers';
 export * from './lib/scrape/PlayerCaptcha';